MT#60078 Add license support for API
API v2 needed to support licenses from /proc/ngcp and restrict access to endpoints based on it.
While the procfiles are yet to be added, an intermediate solution is added. Find a static mapping in license.repository.
When the procfiles are added this need to be revisited and get the licenses from the procfiles.

* Use procfile flags instead of static mapping
* license support for the following endpoints:
  * Pbx
  * Resellers
  * Header Manipulations

Change-Id: I778252b905bf82e4ab04aad4615606086880afb7