MT#61317 Enable strict linting
This is the first step in reshaping all the linting in the application. In summary this adds linting to build process
 installs new linting plugin,  adds new linting rules, updates existing ones, auto fixes auto-fixable problems and manually
fix or ignore others. Ignores can be found by searching for TODO.
Any types and unsafe returns will be resolved in a later commit to be able to focus on them.

* Add linting before build in package.json

* Update existing linting rules
 * indent warn -> error
 * no-console warn -> error
 * comma-dangle warn -> error
 * object-curly-spacing warn -> error
 * array-bracket-spacing warn -> error

* Add new plugin unused-imports
* Add new typescript rules
* no-explicit-any & no-unsafe return set to warning as they need a contained review to fix

Change-Id: I71d9157684fc7c190db2e3d30056bb6850885152