MT#61296 Add import sorting to ESlint
This commit includes a fix for a problem caused by module import reordering.
The order sorting  makes JournalModule gets imported before ExpandModule in a module
that causes circular dependency error in the DI chain of NestJs. It is solved by forwarding the
references correctly. I found no resource on how the DI works in-depth for Nest but it seems the
order of the imports are playing a role in it.

* Imports are now sorted by:
  * group (builtin,internal,parent,sibling,index) with new lines between groups
  * alphabetically
  * multiple imports from the same file are also sorted by name
* Allow ./ import from sameFolder and children directories
* Added new plugin for eslint
  * eslint-plugin-import

Change-Id: I84534a03aa17c8d1ea6a491aa37327c88ec44c2b