MT#61171 Fix explicit any types and unsafe returns
With the latest linting rule update *no-explicit-any* and *no-unsafe-return* were set to warn
to make sure all the errors are contained separately in a review on their own. This commit contains
the fixes after these rules are set to *error* from warn. There are places in the codebase where
its either not possible to fix these because of nestjs limitations or because simply we dont know the type.
These and some other things that needs a second pair of eyes are marked with a *// TODO* comment.

* Add TODO comments for review of unchanged any types & possible questions about their fix
* ServiceRequest is mostly type-safe now, added types to query and headers
* All the functions where sr.query is used should work with the new *QueriesDictionary* type
instead of the wrong ParamsDictionary
* Fix any types and unsafe returns where its possible
* Remove unused message.interface

Change-Id: I4da97f114a0d69c67868f06665c4407212a5074f