MT#61840 Add embedded rules creation for RWR Set
It was requested by the frontend team to be able to create
rwr rules when creating rwr sets. This is a good way to reduce api calls.
Even though this is a solution for this endpoint for now we should aim to
make this reusable for other endpoints in the future.

* New *rewrite-rule-embedded-request* dto to be able to pass rules without set_ids in set request dto
* New *parentId* option for RequestDtoOptions (this will be used for a later implementation heavily)
* Fix typo in 'RewriteRuleResponsetDto'
* Rule Set Service have the option to clear the rules from the sets
* Add test cases to E2E tests

Change-Id: I15ac55efdd39a1cc3410c5c02d89a503cdff3c40